Move entities module one level up
