+++ bids2table/entities.py
